tb: reorganize header files and misc files into subdirectories
