+++ rtl/tb/include/hpdcache_test_amo.h
