+++ rtl/tb/include/hpdcache_test_trace_manager.h
