+++ rtl/tb/sequence_lib/hpdcache_test_exec_trace_seq.h
@@ -17,9 +17,9 @@
  *  under the License.
  */
 /**
- *  Author     : Tommy PRATS 
+ *  Author     : Tommy PRATS
  *  Date       : June, 2025
- *  Description: Class definition of the HPDCACHE test from trace 
+ *  Description: Class definition of the HPDCACHE test from trace
  */
 #ifndef __HPDCACHE_TEST_FROM_FILE_SEQ_H__
 #define __HPDCACHE_TEST_FROM_FILE_SEQ_H__
@@ -31,20 +31,20 @@
 #include "scv.h"
 #include "hpdcache_test_defs.h"
 #include "hpdcache_test_sequence.h"
-#include "hpdcache_trace_manager.h"
+#include "hpdcache_test_trace_manager.h"
 
 #define HPDCACHE_TEST_SEQUENCE_ENABLE_ERROR_SEGMENTS 1
 
 /**
  * @class hpdcache_test_from_trace_seq
- * @brief This class allow to run test by reading transaction directly from a binary trace  
+ * @brief This class allow to run test by reading transaction directly from a binary trace
  *
  */
 class hpdcache_test_exec_trace_seq : public hpdcache_test_sequence
 {
 private:
     typedef sc_bv<HPDCACHE_REQ_DATA_WIDTH> req_data_t;
-    
+
     Trace_reader *my_trace;
     hpdcache_test_sequence::hpdcache_test_memory_segment seg;
 
@@ -94,7 +94,7 @@ public:
 #endif
         int delay_transaction;
         size_t n = 0;
-        while (!my_trace->is_finish() && n < this->max_transactions ) 
+        while (!my_trace->is_finish() && n < this->max_transactions )
         {
             std::shared_ptr<hpdcache_test_transaction_req> t;
             while (!is_available_id()){
@@ -108,7 +108,7 @@ public:
         }
         //  ask the driver to stop
         transaction_fifo_o->write(nullptr);
-        my_trace->my_close(); 
+        my_trace->my_close();
     }
 };
-#endif  // __HPDCACHE_TEST_FROM_FILE_SEQ_H__ 
+#endif  // __HPDCACHE_TEST_FROM_FILE_SEQ_H__
